requirements.txt
antiorm==1.2.1
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
db==0.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
regex==2023.6.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
